requirements/development.txt
-r base.txt

django-extensions==3.0.9
ipython==7.18.1
django-debug-toolbar==3.1.1
